# src/main/java/el/itcompany/controllers/BuildingController.java
package el.itcompany.controllers;

import el.itcompany.entities.Building;
import el.itcompany.entities.Sector;
import el.itcompany.services.BuildingService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/buildings")
public class BuildingController {

    private final BuildingService buildingService;

    public BuildingController(BuildingService buildingService) {
        this.buildingService = buildingService;
    }

    @GetMapping
    public List<Building> getAllBuildings() {
        return buildingService.getAllBuildings();
    }

    @GetMapping("/{id}")
    public ResponseEntity<Building> getBuildingById(@PathVariable Long id) {
        return buildingService.getBuildingById(id)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping
    public Building createBuilding(@RequestBody Building building) {
        return buildingService.createBuilding(building);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Building> updateBuilding(@PathVariable Long id, @RequestBody Building building) {
        try {
            return ResponseEntity.ok(buildingService.updateBuilding(id, building));
        } catch (RuntimeException e) {
            return ResponseEntity.notFound().build();
        }
    }

    @PutMapping("/{id}/addSector")
    public ResponseEntity<Building> addSectorToBuilding(@PathVariable Long id, @RequestBody Sector sector) {
        try {
            return ResponseEntity.ok(buildingService.addSectorToBuilding(id, sector));
        } catch (RuntimeException e) {
            return ResponseEntity.notFound().build();
        }
    }

    @PutMapping("/{id}/removeSector")
    public ResponseEntity<Building> removeSectorFromBuilding(@PathVariable Long id, @RequestBody Sector sector) {
        try {
            return ResponseEntity.ok(buildingService.removeSectorFromBuilding(id, sector));
        } catch (RuntimeException e) {
            return ResponseEntity.notFound().build();
        }
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteBuilding(@PathVariable Long id) {
        buildingService.deleteBuilding(id);
        return ResponseEntity.noContent().build();
    }
}