# swagger.yaml
openapi: 3.1.0
info:
  title: IT-Company - OpenAPI 3.0
  description: This is a sample IT-Company.
  license:
    name: GNU Affero General Public License
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: 3.0.0
  servers:
    - url: 'http://localhost:8080'
...
paths:
...
  /buildings:
    get:
      tags:
        - building
      summary: Get all buildings
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Building'
    post:
      tags:
        - building
      summary: Create a new building
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Building'
      responses:
        '200':
          description: Building created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
  /buildings/{id}:
    get:
      tags:
        - building
      summary: Get building by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Building found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '404':
          description: Building not found
    put:
      tags:
        - building
      summary: Update a building by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Building'
      responses:
        '200':
          description: Building updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '404':
          description: Building not found
    delete:
      tags:
        - building
      summary: Delete building by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Building deleted
...
  /invoices:
    get:
      tags:
        - invoices
      summary: Get all invoices
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
    post:
      tags:
        - invoices
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
  /invoices/{id}:
    get:
      tags:
        - invoices
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          description: Invoice not found
    put:
      tags:
        - invoices
      summary: Update invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
      responses:
        '200':
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          description: Invoice not found
    delete:
      tags:
        - invoices
      summary: Delete invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Invoice deleted
...
  /budget/add:
    post:
      tags:
        - budget
      summary: Add funds to the budget
      parameters:
        - in: query
          name: amount
          required: true
          schema:
            type: number
            format: double
          description: Amount to add
      responses:
        '200':
          description: Updated budget after adding funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyBudget'
  /budget/reserve:
    post:
      summary: Reserve funds from the budget
      tags:
        - budget
      parameters:
        - in: query
          name: amount
          required: true
          schema:
            type: number
            format: double
          description: Amount to reserve
      responses:
        '200':
          description: Updated budget after reserving funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyBudget'
  /budget/debit:
    post:
      summary: Debit reserved funds
      tags:
        - budget
      parameters:
        - in: query
          name: amount
          required: true
          schema:
            type: number
            format: double
          description: Amount to debit from reserved funds
      responses:
        '200':
          description: Updated budget after debiting funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyBudget'
  /budget/release:
    post:
      summary: Release previously reserved funds
      tags:
        - budget
      parameters:
        - in: query
          name: amount
          required: true
          schema:
            type: number
            format: double
          description: Amount to release back into available funds
      responses:
        '200':
          description: Updated budget after releasing funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyBudget'
components:
  schemas:
    Sector:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        salaryClass:
          type: string
          enum: [ A, B, C, D ]
      required:
        - name
        - description
        - salaryClass
    Building:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        city:
          type: string
        capacity:
          type: integer
        sectorList:
          type: array
          items:
            $ref: '#/components/schemas/Sector'
        buildingDate:
          type: string
          format: date
      required:
        - name
        - city
        - capacity
        - buildingDate
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        salary:
          type: integer
        sector:
          $ref: '#/components/schemas/Sector'
        manager:
          type: string
        complaints:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        items:
          type: array
          items:
            type: string
        hireDate:
          type: string
          format: date
      required:
        - firstName
        - lastName
        - email
        - salary
        - hireDate
...
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        manager:
          type: string
        creationDate:
          type: string
          format: date
        workerList:
          type: array
          items:
            type: string
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: integer
        status:
          type: string
          enum: [ PLANNED, ACTIVE, COMPLETED ]
      required:
        - name
        - manager
        - budget
        - status
    Invoice:
      type: object
      properties:
        id:
          type: integer
          format: int64
        client:
          type: string
        project:
          $ref: '#/components/schemas/Project'
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        amount:
          type: number
          format: float
        tax:
          type: number
          format: float
        discount:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        status:
          type: string
          enum: [ DRAFT, SENT, PAID, OVERDUE ]
      required:
        - client
        - issueDate
        - dueDate
        - amount
        - totalAmount
        - status
    CompanyBudget:
      type: object
      properties:
        id:
          type: integer
          format: int64
        totalFunds:
          type: number
          format: double
        reservedFunds:
          type: number
          format: double
        availableFunds:
          type: number
          format: double
      required:
        - totalFunds
        - reservedFunds
        - availableFunds